Generate simulated degradation data

This function is used to generate simulated degradation data.

Usage

sim_dat(group, t, para, process)

Arguments

group

The number of groups of products

t

time.

para

parameters of a certain model.

process

Wiener, Gamma or Inverse Gaussian process.

Value

Return a list containing simulated data at different time points for each group. The output is a list. The first part of the list is the amount of degradation of the simulated data, and the second part is the increment of its degradation

Examples

dat <- sim_dat(group = 6, t = 1:200, para = c(2, 3), process = "Wiener")
head(dat)
#> [[1]]
#>     Time           1           2            3          4          5          6
#> 1      0   0.0000000   0.0000000   0.00000000   0.000000   0.000000   0.000000
#> 2      1   2.5720275  -2.1951044   0.01177294   7.075062   1.453683   3.913765
#> 3      2   1.4841921   0.4387208   0.97217953   8.742784   6.234799   6.939186
#> 4      3   0.2814684   6.5937618   3.09007418  14.253109  15.950850  10.316536
#> 5      4   1.2478124   6.2601311   3.98192937  16.281606  18.830735  22.041037
#> 6      5  -3.4282923   8.5246363  11.37505054  12.895221  22.423215  18.733677
#> 7      6  -2.6039336  10.0978108  16.34756807  12.251203  15.634974  22.055306
#> 8      7   2.8071194  18.2130954  15.99208488  14.477337  12.884275  26.884278
#> 9      8   4.1832967  16.8073292  24.87011473  21.485657  13.977882  32.326265
#> 10     9  11.3361409  21.0953127  27.10717753  21.071472  16.118790  38.462763
#> 11    10  14.4831243  24.6712549  30.54163642  19.392454  21.116920  39.641907
#> 12    11  18.2469062  22.9626252  33.72957562  21.709027  22.735034  46.463113
#> 13    12  18.3508280  24.7197766  34.34950912  28.265209  25.374890  47.731233
#> 14    13  21.4073547  21.3771313  40.25652087  30.693629  23.840401  57.215490
#> 15    14  21.6902410  22.5814818  37.09211130  32.935428  25.225619  63.405298
#> 16    15  23.2268582  19.2700197  40.30650187  34.374687  29.679740  66.316084
#> 17    16  28.3828661  20.3819963  40.89818336  34.025886  34.262637  66.580058
#> 18    17  26.6424995  26.8777929  45.43914364  36.759789  34.968666  69.299043
#> 19    18  32.6550494  31.9046621  45.52314878  32.976344  39.759540  75.207238
#> 20    19  31.4070924  32.7085449  51.23488683  33.451060  43.228325  79.064218
#> 21    20  30.4051722  34.5377995  50.33241993  34.810062  42.180558  84.119778
#> 22    21  30.2478829  33.0278355  51.87420135  31.799903  48.363551  86.151126
#> 23    22  34.9274756  36.2801636  52.63817352  31.393741  53.319152  87.817317
#> 24    23  37.9869778  41.3350986  56.35866613  33.095143  49.724347  87.180668
#> 25    24  37.8442737  42.9439615  62.78676361  35.993477  57.077009  95.633114
#> 26    25  42.8500390  48.5012656  63.42216507  37.467219  62.572686 100.798895
#> 27    26  45.6298307  47.2415634  62.94771086  41.663448  66.361701 106.760455
#> 28    27  45.9713283  45.0958090  69.82303292  45.049603  72.570214 107.668939
#> 29    28  45.6992519  45.1169001  73.61129739  51.415818  71.003630 111.026116
#> 30    29  46.9966379  46.8039956  72.37961909  54.004979  67.342605 111.625331
#> 31    30  49.6077123  45.8524658  68.77244001  55.535745  67.062161 114.428338
#> 32    31  50.5471675  51.5362715  71.09810995  60.157549  69.088814 115.609310
#> 33    32  57.0348451  60.2291337  76.07174126  63.164245  65.616127 111.843329
#> 34    33  60.4640313  59.2841931  80.98692818  67.290915  70.132714 108.891101
#> 35    34  64.3365515  63.7381851  81.50340085  73.330420  73.852547 104.254720
#> 36    35  70.5607581  67.7629629  88.70495823  77.283947  75.939955 108.427669
#> 37    36  69.8627507  75.5898079  94.65129728  81.846076  77.422850 108.990449
#> 38    37  73.9117410  77.2329395  98.58604740  91.392456  74.771399 109.977297
#> 39    38  78.8833518  83.4531287  99.49949784  91.476874  73.058234 111.092932
#> 40    39  82.9453879  85.9254851  99.84512305  93.820074  75.692716 113.619626
#> 41    40  83.0782118  87.5921436  99.84715210  89.197042  79.027717 116.730239
#> 42    41  90.6374737  93.4626187 104.55921813  91.758734  80.569528 121.015469
#> 43    42  98.1209095  98.1726244 103.80135383  87.503070  82.704533 124.304506
#> 44    43 101.0632413 103.6899182 106.76424586  86.495250  87.143784 125.812605
#> 45    44 105.1490329 102.7684524 111.44779591  92.615860  92.788572 128.864042
#> 46    45 106.7698790 102.9512958 110.97779409  97.343658  94.585160 128.036393
#> 47    46 106.0003103 104.3146704 115.74556379  92.243454  91.947416 128.752717
#> 48    47 109.4901727 109.4549646 120.09436247  92.280143  93.538620 132.711646
#> 49    48 112.6355078 106.6637880 116.65519607  97.580917  94.873779 134.391667
#> 50    49 119.2826777 110.4096132 115.87074247  98.875991  99.223269 138.247806
#> 51    50 123.5015331 114.1167297 116.51742245 109.740672 101.461323 139.939365
#> 52    51 125.6903410 114.7176270 119.01923903 112.414808 105.825114 137.691866
#> 53    52 129.1259391 116.6495237 118.74025543 115.651737 113.829991 143.723331
#> 54    53 130.7268104 116.6472864 115.03435261 116.698875 115.113204 152.187372
#> 55    54 129.9155869 125.2394173 109.71936054 118.191240 122.230440 156.707107
#> 56    55 134.4439687 132.1328704 111.02142967 119.616164 123.811582 161.708179
#> 57    56 133.8973667 136.4411319 117.07290717 120.703398 124.342427 164.254610
#> 58    57 135.6479896 134.1366238 116.41245870 124.419884 127.638831 165.268609
#> 59    58 137.5561423 136.5663474 120.41928495 125.727869 131.270482 164.320850
#> 60    59 136.7059754 138.4512391 126.26491776 131.553133 132.497075 165.908925
#> 61    60 140.5746374 143.3189197 136.60009664 132.076878 128.245440 163.877214
#> 62    61 143.7816115 145.6603809 136.67513977 134.882126 133.557821 164.820476
#> 63    62 145.0551194 150.7649555 137.56472016 137.108484 135.008459 164.735611
#> 64    63 150.5008446 155.2474524 137.20838037 139.019271 136.550959 164.281604
#> 65    64 155.3439826 161.2734737 143.14285569 141.119661 141.225968 165.944099
#> 66    65 157.9757578 161.2063949 143.02682202 143.046105 144.335773 167.505964
#> 67    66 160.9217152 163.6083897 145.13857663 140.945264 142.795306 165.393094
#> 68    67 160.5720186 166.7715510 141.50127472 149.723277 143.054884 170.679427
#> 69    68 156.5919071 166.2636456 139.14714486 147.951722 141.599465 176.389544
#> 70    69 162.7159342 168.9128132 142.06491142 153.626001 145.497894 172.811805
#> 71    70 168.7378922 173.3884047 144.53853014 150.693125 144.382540 178.485014
#> 72    71 170.5725844 176.3351571 148.60631817 155.576283 148.528807 176.773234
#> 73    72 173.3345968 183.4545836 150.19546314 157.525752 147.457054 176.262842
#> 74    73 178.7442929 183.2458359 155.00849972 160.637424 152.434588 181.275488
#> 75    74 176.7841079 190.3892429 153.83138826 161.596182 156.107627 183.232701
#> 76    75 175.7760378 192.2780112 161.58206351 158.116579 157.158877 183.490437
#> 77    76 178.1254468 195.5263892 165.49125704 159.536913 158.685268 186.188857
#> 78    77 182.2893755 198.7966000 165.64379550 157.551055 162.359855 186.045855
#> 79    78 179.7672776 202.0382039 171.39564618 154.825652 163.579726 188.524810
#> 80    79 182.6234122 209.7572039 174.00661201 157.322714 166.435704 189.963291
#> 81    80 180.2377121 212.8349249 180.69845816 158.646988 167.950709 195.948228
#> 82    81 179.2856842 217.0372332 191.56963371 166.328339 172.456688 192.063696
#> 83    82 182.3490234 220.7196419 190.81035492 166.270009 174.135098 190.886233
#> 84    83 183.9676577 217.6321850 193.48273962 170.190906 183.650058 190.368460
#> 85    84 183.6192641 224.5426488 191.83604067 173.875795 186.417749 196.831187
#> 86    85 188.5369744 225.5960621 202.07139753 174.596728 183.032028 195.532299
#> 87    86 181.9964309 225.8233624 205.25582319 175.605908 185.144249 200.129206
#> 88    87 187.8036520 231.4179734 205.14308760 176.205101 187.173169 200.133273
#> 89    88 185.9051474 229.9756747 204.83865683 175.125936 186.325217 203.806627
#> 90    89 190.4978024 234.0436082 204.93785648 174.795663 187.748070 200.339789
#> 91    90 194.7020684 238.1854412 202.73821244 175.523935 187.162317 202.751233
#> 92    91 198.2389178 241.7096179 201.88964939 176.005462 191.808620 202.941179
#> 93    92 204.4735334 249.2028255 204.74691633 180.345513 196.083979 209.657234
#> 94    93 205.0375649 249.6798294 204.96238896 180.803068 201.147499 217.929591
#> 95    94 205.8189461 253.8283572 205.73097873 188.110211 198.844884 218.960660
#> 96    95 206.8917979 260.1790534 208.30035831 193.584652 202.445572 221.404914
#> 97    96 212.7486551 264.8913369 208.18462413 194.766423 211.224390 221.653894
#> 98    97 215.1569924 263.2044040 204.41294953 198.258833 207.843011 224.537486
#> 99    98 221.3846400 265.9201536 209.30731929 200.403730 211.599396 220.060228
#> 100   99 224.9075097 264.9319452 210.57534269 196.727482 213.861971 224.698626
#> 101  100 230.4264097 270.0087168 210.74079543 199.996871 212.771523 229.642854
#> 102  101 232.0536250 274.1146318 212.96738064 206.156737 210.976674 231.298835
#> 103  102 232.5530358 272.7659357 213.51612491 206.130603 212.477572 233.556754
#> 104  103 230.6761138 268.2663472 216.96301709 207.550986 220.572639 238.483827
#> 105  104 231.3684959 268.6410135 218.26286723 207.056030 222.501631 240.474523
#> 106  105 229.1250500 271.4393355 222.90338332 207.703680 224.997841 245.539825
#> 107  106 228.2732484 274.8976605 225.32481627 209.579860 226.277194 242.584463
#> 108  107 230.1776886 275.4388190 230.11296947 212.151107 228.582708 243.078179
#> 109  108 228.3502114 274.9732963 231.66758422 213.877394 229.733439 247.784511
#> 110  109 228.3149214 278.8727034 235.95065761 219.027947 231.842987 251.444934
#> 111  110 230.7422784 277.9479748 234.94039644 218.993463 232.303988 252.278648
#> 112  111 233.3340581 280.8771674 235.16842180 220.327573 239.125862 251.396655
#> 113  112 239.1055135 283.6858842 233.37077761 226.768686 237.976118 254.818020
#> 114  113 245.5890573 279.0788338 229.74513092 226.281420 242.807191 255.608199
#> 115  114 250.6282498 275.5055662 232.55129154 230.833590 244.038299 255.149847
#> 116  115 252.0560520 277.3128208 236.82092724 233.342953 244.963746 258.213177
#> 117  116 251.0923566 279.4307326 239.85321916 233.685219 250.615901 254.283049
#> 118  117 253.3371231 282.7025091 243.12059310 237.124971 248.312635 253.326898
#> 119  118 254.2417385 292.6463995 247.20801444 236.186668 250.708945 253.296583
#> 120  119 253.7778380 294.9191574 247.96999587 234.682097 258.269876 256.832046
#> 121  120 255.5577999 294.8324080 250.05958819 234.997747 260.520055 261.797417
#> 122  121 258.0395042 294.2199424 251.77324520 234.164754 265.643386 265.689376
#> 123  122 263.6876627 298.5818570 251.22083128 237.494140 268.009923 265.982698
#> 124  123 268.0714842 299.1937403 258.32080687 236.563562 272.019154 265.486243
#> 125  124 269.5976339 298.7660382 258.46755883 233.911500 271.780281 269.499836
#> 126  125 271.0023207 298.8956108 259.30090523 235.225989 278.328533 273.148975
#> 127  126 270.8337055 298.7632111 265.00335175 238.902071 281.646740 274.037232
#> 128  127 274.2624980 301.3191754 262.49765320 241.933211 285.902112 274.828728
#> 129  128 279.8996698 302.4808861 267.86574487 244.298429 287.517814 276.222224
#> 130  129 280.8201473 306.0273308 272.75442372 243.161451 286.211819 283.662814
#> 131  130 284.5184037 310.8796386 270.86596054 247.728362 290.490542 283.374018
#> 132  131 284.0399585 313.0603945 277.17452315 251.774759 291.959727 283.559228
#> 133  132 287.1422632 312.2530542 277.28176179 250.543404 300.510541 281.307893
#> 134  133 289.6227746 317.0750958 280.18217860 248.722921 302.262957 284.137722
#> 135  134 292.1819212 319.8924903 281.23844863 255.869471 308.278701 289.634126
#> 136  135 291.2164176 321.0858095 285.17108688 253.958046 309.481182 294.843737
#> 137  136 293.7961209 325.4949723 288.30172168 259.041652 312.546884 300.791402
#> 138  137 297.9469795 326.7292955 298.03781403 265.699990 317.658450 303.194627
#> 139  138 300.0013337 335.0309575 300.56683942 258.605068 315.435654 306.863351
#> 140  139 299.8655431 332.8472741 306.59120011 261.892515 317.303449 310.490202
#> 141  140 301.0089196 341.0165644 307.78259414 267.651728 318.128207 313.452687
#> 142  141 303.7296916 340.1457532 305.76349356 266.398035 320.243859 317.978644
#> 143  142 300.9437899 338.2463516 308.30368835 270.064534 318.282376 323.524308
#> 144  143 300.8727873 341.2602074 312.02259562 274.122778 321.554128 331.383731
#> 145  144 305.5828229 340.5679969 319.20266225 272.177553 325.360289 327.951336
#> 146  145 307.4342271 341.5197249 322.57499351 272.851149 325.932097 326.783808
#> 147  146 311.6769683 347.4241668 324.45205727 269.643853 328.247626 328.690285
#> 148  147 311.1644134 348.4342339 322.04359919 273.257272 332.150129 329.706331
#> 149  148 308.9480536 344.2182126 326.23823292 280.717307 329.975229 330.911978
#> 150  149 308.6515464 343.1772007 326.87149471 281.785427 329.570162 338.002414
#> 151  150 305.0284432 342.2548962 330.78729276 283.865227 332.441851 338.540668
#> 152  151 306.8865359 343.7492938 325.95939317 288.486115 335.030451 335.737385
#> 153  152 309.6115121 348.2948311 322.47925842 290.383131 335.962232 337.096065
#> 154  153 308.4577223 349.9097862 318.82597104 295.096110 340.283187 340.059386
#> 155  154 312.1630406 354.1258276 315.67392214 295.178493 341.354919 343.568400
#> 156  155 320.2644452 356.6577533 317.77834973 298.980621 341.607662 345.094195
#> 157  156 323.4146681 357.2126696 321.76376156 302.011695 349.689813 348.162758
#> 158  157 329.0462288 356.2576262 327.07867654 301.909159 346.388660 345.178883
#> 159  158 330.4666503 358.4533421 325.25698647 301.436984 351.179424 348.531759
#> 160  159 331.0756028 362.2421700 325.98515563 306.784268 354.500349 349.788516
#> 161  160 330.2444128 364.3092542 327.76084785 303.792880 356.859599 350.812615
#> 162  161 329.5400078 366.0975148 331.38567200 304.892964 358.794310 353.160957
#> 163  162 330.3170654 369.0898837 331.59183830 303.542144 362.763992 351.967334
#> 164  163 328.3692096 371.2806896 334.53836350 304.490021 366.998809 353.913204
#> 165  164 328.2832119 373.3440484 328.22739098 308.971904 371.316436 358.742578
#> 166  165 326.3377671 373.5079890 330.59401842 310.713458 371.051372 360.416036
#> 167  166 326.7284513 377.4818606 336.04578468 311.955310 376.139999 364.296376
#> 168  167 336.4930641 377.0798632 341.81148123 316.529244 375.023441 364.654976
#> 169  168 335.1968804 379.1879242 344.26202484 317.981941 378.836267 365.186968
#> 170  169 338.8927783 379.1401276 346.52711499 322.296383 373.776037 364.163424
#> 171  170 344.9981175 377.6093776 349.82273419 326.214670 376.424276 362.782540
#> 172  171 342.7313943 377.7561777 350.96663101 326.191748 378.519569 365.175548
#> 173  172 345.4799234 377.7158837 350.67901148 322.552954 380.571551 367.037234
#> 174  173 342.7991175 377.1825304 352.28233883 319.808680 382.793084 367.265849
#> 175  174 338.7622030 383.6655754 354.74637467 318.550742 388.879708 372.595976
#> 176  175 337.9649073 392.9245685 354.48412028 318.900860 394.035253 372.911779
#> 177  176 340.1113591 394.3991783 350.83969825 321.353371 395.290130 376.286537
#> 178  177 345.6812541 403.8154073 354.44321212 327.865281 392.188153 379.146602
#> 179  178 350.5430568 403.9641431 351.59242706 329.367569 398.307826 385.156501
#> 180  179 351.7082772 405.7123314 358.78159143 326.605455 395.749317 388.124514
#> 181  180 354.2127530 405.3282993 360.40519876 329.785287 402.020118 387.375042
#> 182  181 361.9689150 404.4841817 364.68029847 328.668795 398.492646 391.105241
#> 183  182 365.3785915 406.3614658 365.34135123 331.507061 405.696163 389.791329
#> 184  183 364.7167406 408.6209522 366.97683305 333.263821 409.603180 390.568490
#> 185  184 367.6432012 407.3691562 365.11266103 337.488082 409.255406 389.735399
#> 186  185 369.8171974 410.3160061 367.49741067 341.661587 413.750574 392.593240
#> 187  186 370.4700018 408.4147551 369.69157654 344.779772 410.950294 393.813088
#> 188  187 371.8997271 412.0339182 369.58322477 341.733054 412.787065 392.988637
#> 189  188 371.8488765 414.8001584 370.41815723 342.212571 418.807807 395.512763
#> 190  189 370.6729817 413.8866399 372.28526310 347.338321 417.733685 393.808011
#> 191  190 370.0045268 416.4200895 373.85187911 346.819982 422.423234 398.401788
#> 192  191 372.7111471 416.7805947 378.12066440 343.055945 427.480725 394.075095
#> 193  192 376.2074860 414.4670527 378.40779200 347.671004 428.892182 398.639632
#> 194  193 382.1010767 420.7531833 381.51389781 349.144026 427.358448 409.525380
#> 195  194 382.3320575 421.3232336 379.50985636 351.147635 429.411939 405.259514
#> 196  195 385.0253959 432.5437009 384.74227103 353.537943 429.505658 407.067062
#> 197  196 388.3046610 437.9382266 384.84474117 355.607813 431.497924 409.543024
#> 198  197 387.6105397 442.0486706 386.04083016 358.272813 437.287152 410.970883
#> 199  198 390.8153752 440.4245944 389.47036245 359.837571 439.651879 413.486398
#> 200  199 396.3780061 446.6220381 387.34290538 365.118451 446.827985 417.108204
#> 
#> [[2]]
#>     Time           1            2            3           4           5
#> 1      0  0.00000000  0.000000000  0.000000000  0.00000000  0.00000000
#> 2      1  2.57202752 -2.195104439  0.011772937  7.07506152  1.45368251
#> 3      2 -1.08783538  2.633825268  0.960406593  1.66772290  4.78111656
#> 4      3 -1.20272374  6.155040931  2.117894653  5.51032447  9.71605084
#> 5      4  0.96634404 -0.333630639  0.891855182  2.02849750  2.87988518
#> 6      5 -4.67610473  2.264505197  7.393121178 -3.38638580  3.59247948
#> 7      6  0.82435867  1.573174467  4.972517523 -0.64401726 -6.78824074
#> 8      7  5.41105307  8.115284595 -0.355483188  2.22613385 -2.75069852
#> 9      8  1.37617728 -1.405766201  8.878029855  7.00831965  1.09360620
#> 10     9  7.15284415  4.287983530  2.237062797 -0.41418489  2.14090852
#> 11    10  3.14698344  3.575942239  3.434458888 -1.67901799  4.99812953
#> 12    11  3.76378184 -1.708629753  3.187939207  2.31657315  1.61811429
#> 13    12  0.10392187  1.757151377  0.619933495  6.55618182  2.63985644
#> 14    13  3.05652669 -3.342645243  5.907011752  2.42842054 -1.53448937
#> 15    14  0.28288623  1.204350513 -3.164409577  2.24179823  1.38521787
#> 16    15  1.53661721 -3.311462119  3.214390571  1.43925886  4.45412136
#> 17    16  5.15600790  1.111976534  0.591681496 -0.34880048  4.58289657
#> 18    17 -1.74036658  6.495796594  4.540960282  2.73390267  0.70602907
#> 19    18  6.01254989  5.026869218  0.084005132 -3.78344479  4.79087437
#> 20    19 -1.24795700  0.803882786  5.711738052  0.47471593  3.46878487
#> 21    20 -1.00192018  1.829254600 -0.902466896  1.35900249 -1.04776686
#> 22    21 -0.15728930 -1.509963939  1.541781417 -3.01015960  6.18299249
#> 23    22  4.67959273  3.252328038  0.763972172 -0.40616162  4.95560151
#> 24    23  3.05950214  5.054935036  3.720492611  1.70140171 -3.59480560
#> 25    24 -0.14270408  1.608862920  6.428097478  2.89833452  7.35266265
#> 26    25  5.00576530  5.557304109  0.635401461  1.47374212  5.49567715
#> 27    26  2.77979169 -1.259702233 -0.474454216  4.19622819  3.78901457
#> 28    27  0.34149764 -2.145754364  6.875322069  3.38615490  6.20851275
#> 29    28 -0.27207644  0.021091084  3.788264464  6.36621535 -1.56658363
#> 30    29  1.29738599  1.687095487 -1.231678301  2.58916060 -3.66102511
#> 31    30  2.61107443 -0.951529756 -3.607179081  1.53076605 -0.28044430
#> 32    31  0.93945521  5.683805630  2.325669938  4.62180408  2.02665341
#> 33    32  6.48767756  8.692862220  4.973631320  3.00669634 -3.47268700
#> 34    33  3.42918622 -0.944940538  4.915186910  4.12666993  4.51658738
#> 35    34  3.87252022  4.453991982  0.516472677  6.03950494  3.71983290
#> 36    35  6.22420657  4.024777781  7.201557379  3.95352678  2.08740730
#> 37    36 -0.69800741  7.826845012  5.946339049  4.56212949  1.48289511
#> 38    37  4.04899036  1.643131618  3.934750117  9.54637943 -2.65145106
#> 39    38  4.97161076  6.220189124  0.913450442  0.08441826 -1.71316445
#> 40    39  4.06203615  2.472356449  0.345625210  2.34320050  2.63448156
#> 41    40  0.13282390  1.666658533  0.002029053 -4.62303249  3.33500073
#> 42    41  7.55926187  5.870475043  4.712066024  2.56169208  1.54181098
#> 43    42  7.48343580  4.710005662 -0.757864291 -4.25566387  2.13500583
#> 44    43  2.94233179  5.517293881  2.962892026 -1.00781964  4.43925105
#> 45    44  4.08579166 -0.921465820  4.683550050  6.12060948  5.64478730
#> 46    45  1.62084607  0.182843356 -0.470001825  4.72779779  1.79658832
#> 47    46 -0.76956870  1.363374608  4.767769708 -5.10020359 -2.63774389
#> 48    47  3.48986243  5.140294269  4.348798676  0.03668915  1.59120349
#> 49    48  3.14533507 -2.791176624 -3.439166398  5.30077352  1.33515970
#> 50    49  6.64716989  3.745825186 -0.784453600  1.29507417  4.34949013
#> 51    50  4.21885538  3.707116519  0.646679982 10.86468145  2.23805316
#> 52    51  2.18880791  0.600897302  2.501816574  2.67413543  4.36379161
#> 53    52  3.43559811  1.931896682 -0.278983595  3.23692888  8.00487645
#> 54    53  1.60087129 -0.002237314 -3.705902828  1.04713858  1.28321353
#> 55    54 -0.81122347  8.592130939 -5.314992065  1.49236453  7.11723595
#> 56    55  4.52838179  6.893453092  1.302069125  1.42492451  1.58114182
#> 57    56 -0.54660206  4.308261460  6.051477500  1.08723406  0.53084518
#> 58    57  1.75062290 -2.304508093 -0.660448462  3.71648578  3.29640412
#> 59    58  1.90815275  2.429723605  4.006826246  1.30798439  3.63165081
#> 60    59 -0.85016696  1.884891655  5.845632809  5.82526422  1.22659323
#> 61    60  3.86866209  4.867680595 10.335178885  0.52374551 -4.25163506
#> 62    61  3.20697403  2.341461225  0.075043130  2.80524735  5.31238067
#> 63    62  1.27350790  5.104574646  0.889580388  2.22635800  1.45063840
#> 64    63  5.44572527  4.482496862 -0.356339797  1.91078741  1.54250008
#> 65    64  4.84313796  6.026021273  5.934475329  2.10038954  4.67500847
#> 66    65  2.63177519 -0.067078802 -0.116033675  1.92644447  3.10980536
#> 67    66  2.94595745  2.401994880  2.111754614 -2.10084149 -1.54046740
#> 68    67 -0.34969666  3.163161252 -3.637301916  8.77801361  0.25957808
#> 69    68 -3.98011144 -0.507905353 -2.354129852 -1.77155505 -1.45541930
#> 70    69  6.12402708  2.649167525  2.917766551  5.67427853  3.89842969
#> 71    70  6.02195797  4.475591579  2.473618720 -2.93287567 -1.11535420
#> 72    71  1.83469223  2.946752326  4.067788036  4.88315770  4.14626691
#> 73    72  2.76201235  7.119426539  1.589144964  1.94946923 -1.07175300
#> 74    73  5.40969616 -0.208747709  4.813036586  3.11167233  4.97753379
#> 75    74 -1.96018501  7.143407050 -1.177111464  0.95875751  3.67303875
#> 76    75 -1.00807015  1.888768244  7.750675252 -3.47960237  1.05125093
#> 77    76  2.34940900  3.248377969  3.909193527  1.42033347  1.52639057
#> 78    77  4.16392877  3.270210878  0.152538461 -1.98585759  3.67458721
#> 79    78 -2.52209798  3.241603856  5.751850684 -2.72540318  1.21987027
#> 80    79  2.85613466  7.719000032  2.610965824  2.49706235  2.85597813
#> 81    80 -2.38570008  3.077720999  6.691846154  1.32427353  1.51500505
#> 82    81 -0.95202798  4.202308233 10.871175545  7.68135092  4.50597961
#> 83    82  3.06333922  3.682408777 -0.759278784 -0.05833026  1.67840955
#> 84    83  1.61863429 -3.087456943  2.672384699  3.92089739  9.51495991
#> 85    84 -0.34839355  6.910463821 -1.646698948  3.68488874  2.76769164
#> 86    85  4.91771028  1.053413292 10.235356854  0.72093315 -3.38572182
#> 87    86 -6.54054353  0.227300334  3.184425663  1.00917981  2.11222166
#> 88    87  5.80722114  5.594610993 -0.112735593  0.59919304  2.02891941
#> 89    88 -1.89850464 -1.442298726 -0.304430764 -1.07916455 -0.84795155
#> 90    89  4.59265500  4.067933465  0.099199652 -0.33027286  1.42285318
#> 91    90  4.20426599  4.141833052 -2.199644043  0.72827171 -0.58575305
#> 92    91  3.53684939  3.524176690 -0.848563047  0.48152682  4.64630264
#> 93    92  6.23461563  7.493207551  2.857266932  4.34005085  4.27535906
#> 94    93  0.56403151  0.477003933  0.215472637  0.45755552  5.06352021
#> 95    94  0.78138120  4.148527798  0.768589764  7.30714241 -2.30261531
#> 96    95  1.07285180  6.350696192  2.569379582  5.47444156  3.60068827
#> 97    96  5.85685721  4.712283501 -0.115734180  1.18177063  8.77881809
#> 98    97  2.40833729 -1.686932885 -3.771674602  3.49241001 -3.38137931
#> 99    98  6.22764756  2.715749605  4.894369765  2.14489697  3.75638531
#> 100   99  3.52286978 -0.988208378  1.268023402 -3.67624743  2.26257474
#> 101  100  5.51889996  5.076771599  0.165452733  3.26938896 -1.09044816
#> 102  101  1.62721535  4.105914975  2.226585213  6.15986564 -1.79484884
#> 103  102  0.49941075 -1.348696147  0.548744265 -0.02613394  1.50089817
#> 104  103 -1.87692195 -4.499588408  3.446892186  1.42038339  8.09506641
#> 105  104  0.69238210  0.374666208  1.299850136 -0.49495641  1.92899254
#> 106  105 -2.24344595  2.798322072  4.640516090  0.64765022  2.49620992
#> 107  106 -0.85180155  3.458324936  2.421432951  1.87617977  1.27935255
#> 108  107  1.90444012  0.541158582  4.788153196  2.57124740  2.30551438
#> 109  108 -1.82747716 -0.465522792  1.554614752  1.72628672  1.15073110
#> 110  109 -0.03529004  3.899407138  4.283073390  5.15055257  2.10954808
#> 111  110  2.42735702 -0.924728570 -1.010261166 -0.03448381  0.46100118
#> 112  111  2.59177975  2.929192568  0.228025362  1.33410975  6.82187391
#> 113  112  5.77145536  2.808716792 -1.797644198  6.44111373 -1.14974448
#> 114  113  6.48354382 -4.607050409 -3.625646688 -0.48726651  4.83107334
#> 115  114  5.03919247 -3.573267576  2.806160622  4.55216963  1.23110838
#> 116  115  1.42780220  1.807254593  4.269635703  2.50936356  0.92544619
#> 117  116 -0.96369537  2.117911810  3.032291915  0.34226612  5.65215531
#> 118  117  2.24476654  3.271776524  3.267373947  3.43975189 -2.30326636
#> 119  118  0.90461539  9.943890336  4.087421331 -0.93830357  2.39631061
#> 120  119 -0.46390053  2.272757970  0.761981437 -1.50457067  7.56093077
#> 121  120  1.77996190 -0.086749466  2.089592322  0.31565021  2.25017951
#> 122  121  2.48170428 -0.612465614  1.713657007 -0.83299341  5.12333083
#> 123  122  5.64815849  4.361914644 -0.552413920  3.32938677  2.36653699
#> 124  123  4.38382155  0.611883341  7.099975587 -0.93057817  4.00923114
#> 125  124  1.52614968 -0.427702119  0.146751959 -2.65206206 -0.23887361
#> 126  125  1.40468682  0.129572542  0.833346402  1.31448897  6.54825254
#> 127  126 -0.16861519 -0.132399630  5.702446518  3.67608147  3.31820652
#> 128  127  3.42879250  2.555964288 -2.505698547  3.03114051  4.25537258
#> 129  128  5.63717176  1.161710729  5.368091669  2.36521836  1.61570180
#> 130  129  0.92047749  3.546444655  4.888678854 -1.13697804 -1.30599527
#> 131  130  3.69825639  4.852307762 -1.888463186  4.56691017  4.27872295
#> 132  131 -0.47844521  2.180755912  6.308562616  4.04639768  1.46918506
#> 133  132  3.10230475 -0.807340260  0.107238640 -1.23135569  8.55081398
#> 134  133  2.48051137  4.822041579  2.900416807 -1.82048297  1.75241642
#> 135  134  2.55914659  2.817394526  1.056270034  7.14655086  6.01574324
#> 136  135 -0.96550354  1.193319133  3.932638245 -1.91142568  1.20248099
#> 137  136  2.57970327  4.409162877  3.130634803  5.08360631  3.06570225
#> 138  137  4.15085858  1.234323162  9.736092345  6.65833766  5.11156567
#> 139  138  2.05435427  8.301661976  2.529025391 -7.09492160 -2.22279505
#> 140  139 -0.13579065 -2.183683386  6.024360696  3.28744732  1.86779485
#> 141  140  1.14337655  8.169290350  1.191394024  5.75921236  0.82475734
#> 142  141  2.72077191 -0.870811255 -2.019100583 -1.25369267  2.11565263
#> 143  142 -2.78590162 -1.899401531  2.540194795  3.66649871 -1.96148369
#> 144  143 -0.07100262  3.013855759  3.718907266  4.05824376  3.27175286
#> 145  144  4.71003561 -0.692210501  7.180066632 -1.94522498  3.80616079
#> 146  145  1.85140418  0.951727988  3.372331259  0.67359599  0.57180749
#> 147  146  4.24274116  5.904441925  1.877063766 -3.20729612  2.31552879
#> 148  147 -0.51255489  1.010067068 -2.408458083  3.61341962  3.90250372
#> 149  148 -2.21635977 -4.216021319  4.194633726  7.46003513 -2.17490041
#> 150  149 -0.29650719 -1.041011883  0.633261794  1.06811958 -0.40506663
#> 151  150 -3.62310322 -0.922304522  3.915798045  2.07980016  2.87168892
#> 152  151  1.85809272  1.494397630 -4.827899591  4.62088811  2.58860003
#> 153  152  2.72497624  4.545537328 -3.480134747  1.89701574  0.93178038
#> 154  153 -1.15378988  1.614955097 -3.653287379  4.71297898  4.32095592
#> 155  154  3.70531835  4.216041361 -3.152048897  0.08238309  1.07173167
#> 156  155  8.10140462  2.531925706  2.104427591  3.80212804  0.25274333
#> 157  156  3.15022285  0.554916366  3.985411824  3.03107400  8.08215059
#> 158  157  5.63156070 -0.955043436  5.314914986 -0.10253630 -3.30115352
#> 159  158  1.42042151  2.195715934 -1.821690071 -0.47217505  4.79076419
#> 160  159  0.60895251  3.788827887  0.728169158  5.34728440  3.32092492
#> 161  160 -0.83118997  2.067084140  1.775692219 -2.99138774  2.35925023
#> 162  161 -0.70440506  1.788260656  3.624824149  1.10008403  1.93471110
#> 163  162  0.77705762  2.992368847  0.206166299 -1.35082061  3.96968183
#> 164  163 -1.94785582  2.190805921  2.946525201  0.94787716  4.23481688
#> 165  164 -0.08599765  2.063358778 -6.310972523  4.48188329  4.31762748
#> 166  165 -1.94544487  0.163940599  2.366627439  1.74155346 -0.26506411
#> 167  166  0.39068427  3.973871599  5.451766265  1.24185214  5.08862697
#> 168  167  9.76461275 -0.401997399  5.765696545  4.57393445 -1.11655854
#> 169  168 -1.29618368  2.108061008  2.450543610  1.45269671  3.81282681
#> 170  169  3.69589792 -0.047796566  2.265090152  4.31444188 -5.06023033
#> 171  170  6.10533920 -1.530750063  3.295619200  3.91828697  2.64823949
#> 172  171 -2.26672318  0.146800122  1.143896827 -0.02292126  2.09529228
#> 173  172  2.74852909 -0.040294008 -0.287619535 -3.63879425  2.05198223
#> 174  173 -2.68080590 -0.533353233  1.603327352 -2.74427420  2.22153288
#> 175  174 -4.03691455  6.483045016  2.464035838 -1.25793839  6.08662430
#> 176  175 -0.79729564  9.258993017 -0.262254392  0.35011790  5.15554479
#> 177  176  2.14645177  1.474609883 -3.644422023  2.45251160  1.25487687
#> 178  177  5.56989504  9.416228918  3.603513864  6.51190953 -3.10197699
#> 179  178  4.86180264  0.148735822 -2.850785061  1.50228861  6.11967270
#> 180  179  1.16522037  1.748188313  7.189164374 -2.76211392 -2.55850816
#> 181  180  2.50447579 -0.384032103  1.623607324  3.17983205  6.27080096
#> 182  181  7.75616204 -0.844117565  4.275099717 -1.11649245 -3.52747193
#> 183  182  3.40967653  1.877284018  0.661052760  2.83826623  7.20351694
#> 184  183 -0.66185094  2.259486439  1.635481818  1.75675967  3.90701642
#> 185  184  2.92646057 -1.251795974 -1.864172024  4.22426109 -0.34777369
#> 186  185  2.17399627  2.946849871  2.384749643  4.17350497  4.49516791
#> 187  186  0.65280439 -1.901251021  2.194165871  3.11818499 -2.80027977
#> 188  187  1.42972524  3.619163148 -0.108351772 -3.04671782  1.83677040
#> 189  188 -0.05085053  2.766240182  0.834932461  0.47951689  6.02074215
#> 190  189 -1.17589486 -0.913518452  1.867105876  5.12574991 -1.07412205
#> 191  190 -0.66845490  2.533449547  1.566616001 -0.51833881  4.68954962
#> 192  191  2.70662034  0.360505211  4.268785293 -3.76403716  5.05749060
#> 193  192  3.49633886 -2.313542002  0.287127600  4.61505954  1.41145712
#> 194  193  5.89359074  6.286130642  3.106105811  1.47302170 -1.53373388
#> 195  194  0.23098079  0.570050260 -2.004041445  2.00360895  2.05349047
#> 196  195  2.69333842 11.220467273  5.232414664  2.39030816  0.09371918
#> 197  196  3.27926513  5.394525733  0.102470142  2.06986962  1.99226621
#> 198  197 -0.69412133  4.110444007  1.196088987  2.66499985  5.78922821
#> 199  198  3.20483553 -1.624076250  3.429532297  1.56475812  2.36472649
#> 200  199  5.56263083  6.197443777 -2.127457075  5.28088017  7.17610579
#>                6
#> 1    0.000000000
#> 2    3.913765318
#> 3    3.025420681
#> 4    3.377349617
#> 5   11.724501046
#> 6   -3.307359449
#> 7    3.321628689
#> 8    4.828971757
#> 9    5.441987664
#> 10   6.136497811
#> 11   1.179144124
#> 12   6.821206136
#> 13   1.268119402
#> 14   9.484257648
#> 15   6.189807415
#> 16   2.910786364
#> 17   0.263973849
#> 18   2.718985421
#> 19   5.908194429
#> 20   3.856980415
#> 21   5.055559483
#> 22   2.031348522
#> 23   1.666190411
#> 24  -0.636648403
#> 25   8.452446027
#> 26   5.165780299
#> 27   5.961560364
#> 28   0.908484329
#> 29   3.357177041
#> 30   0.599214729
#> 31   2.803006442
#> 32   1.180972769
#> 33  -3.765981290
#> 34  -2.952228175
#> 35  -4.636380749
#> 36   4.172949143
#> 37   0.562779428
#> 38   0.986847947
#> 39   1.115635064
#> 40   2.526693822
#> 41   3.110613569
#> 42   4.285230172
#> 43   3.289036237
#> 44   1.508099777
#> 45   3.051436279
#> 46  -0.827648867
#> 47   0.716323874
#> 48   3.958929602
#> 49   1.680020963
#> 50   3.856139091
#> 51   1.691558480
#> 52  -2.247498559
#> 53   6.031464831
#> 54   8.464040545
#> 55   4.519735910
#> 56   5.001071678
#> 57   2.546431253
#> 58   1.013998588
#> 59  -0.947758584
#> 60   1.588074609
#> 61  -2.031711007
#> 62   0.943261681
#> 63  -0.084864754
#> 64  -0.454007394
#> 65   1.662495913
#> 66   1.561864864
#> 67  -2.112870644
#> 68   5.286333432
#> 69   5.710116823
#> 70  -3.577738586
#> 71   5.673208568
#> 72  -1.711780172
#> 73  -0.510391621
#> 74   5.012645630
#> 75   1.957213721
#> 76   0.257735432
#> 77   2.698420377
#> 78  -0.143001776
#> 79   2.478954305
#> 80   1.438481343
#> 81   5.984936504
#> 82  -3.884531605
#> 83  -1.177463450
#> 84  -0.517772305
#> 85   6.462726407
#> 86  -1.298887445
#> 87   4.596907190
#> 88   0.004066378
#> 89   3.673354189
#> 90  -3.466837578
#> 91   2.411443242
#> 92   0.189946577
#> 93   6.716054976
#> 94   8.272356444
#> 95   1.031068989
#> 96   2.444254040
#> 97   0.248980042
#> 98   2.883592128
#> 99  -4.477257872
#> 100  4.638397541
#> 101  4.944228154
#> 102  1.655981579
#> 103  2.257918302
#> 104  4.927073394
#> 105  1.990696379
#> 106  5.065302088
#> 107 -2.955362675
#> 108  0.493716050
#> 109  4.706332633
#> 110  3.660422471
#> 111  0.833714190
#> 112 -0.881992714
#> 113  3.421364204
#> 114  0.790179481
#> 115 -0.458351981
#> 116  3.063329594
#> 117 -3.930127394
#> 118 -0.956151642
#> 119 -0.030314218
#> 120  3.535463043
#> 121  4.965370412
#> 122  3.891959286
#> 123  0.293321970
#> 124 -0.496454969
#> 125  4.013593247
#> 126  3.649138318
#> 127  0.888257137
#> 128  0.791495846
#> 129  1.393496318
#> 130  7.440590046
#> 131 -0.288795612
#> 132  0.185210045
#> 133 -2.251335889
#> 134  2.829829026
#> 135  5.496404273
#> 136  5.209610684
#> 137  5.947665217
#> 138  2.403225565
#> 139  3.668723198
#> 140  3.626851554
#> 141  2.962485325
#> 142  4.525956266
#> 143  5.545664419
#> 144  7.859422580
#> 145 -3.432394540
#> 146 -1.167528054
#> 147  1.906476514
#> 148  1.016046385
#> 149  1.205647245
#> 150  7.090436196
#> 151  0.538253906
#> 152 -2.803283216
#> 153  1.358679429
#> 154  2.963321372
#> 155  3.509014322
#> 156  1.525794808
#> 157  3.068562560
#> 158 -2.983874501
#> 159  3.352876292
#> 160  1.256757037
#> 161  1.024098195
#> 162  2.348341834
#> 163 -1.193622451
#> 164  1.945869776
#> 165  4.829373739
#> 166  1.673458091
#> 167  3.880340029
#> 168  0.358600363
#> 169  0.531991465
#> 170 -1.023543667
#> 171 -1.380884218
#> 172  2.393008328
#> 173  1.861686199
#> 174  0.228615029
#> 175  5.330127170
#> 176  0.315802855
#> 177  3.374757332
#> 178  2.860065339
#> 179  6.009899510
#> 180  2.968012894
#> 181 -0.749472485
#> 182  3.730199051
#> 183 -1.313911534
#> 184  0.777160452
#> 185 -0.833090480
#> 186  2.857840225
#> 187  1.219848298
#> 188 -0.824451222
#> 189  2.524126397
#> 190 -1.704752272
#> 191  4.593776976
#> 192 -4.326692348
#> 193  4.564536528
#> 194 10.885747771
#> 195 -4.265865838
#> 196  1.807548369
#> 197  2.475962241
#> 198  1.427858139
#> 199  2.515514947
#> 200  3.621806159
#>