Generate simulated degradation data

This function is used to generate simulated degradation data.

Usage

sim_dat(group, t, para, process)

Arguments

group

The number of groups of products

t

time.

para

parameters of a certain model.

process

Wiener, Gamma or Inverse Gaussian process.

Value

Return a list containing simulated data at different time points for each group. The output is a list. The first part of the list is the amount of degradation of the simulated data, and the second part is the increment of its degradation

Examples

dat <- sim_dat(group = 6, t = 1:200, para = c(2, 3), process = "Wiener")
head(dat)
#> [[1]]
#>     Time          1          2           3          4           5           6
#> 1      0   0.000000   0.000000   0.0000000   0.000000   0.0000000   0.0000000
#> 2      1   4.722417  -3.564218   0.6412322   4.839179  -1.0722159   2.6159263
#> 3      2  12.435652  -4.022833   3.7856642   7.026717  -0.5037355   0.8909609
#> 4      3  14.539512  -4.218012   5.5232622  10.546678   0.6002263  -0.9031893
#> 5      4  12.790515  -3.918971   5.3662256  16.903253   3.1734953  -5.8852504
#> 6      5  17.053547  -5.792074   7.7330392  16.813756   5.9464285  -3.1279450
#> 7      6  15.228547  -1.548647   9.9304534  15.487257   5.6657649  -1.6689520
#> 8      7  19.305371   1.266003  13.2440628  19.399697   6.1128592  -1.2851313
#> 9      8  20.294194   6.208114  15.1448714  19.847130   5.3939779   0.6707439
#> 10     9  21.166077   9.459228  16.9345445  28.083427   0.4057144   3.1494055
#> 11    10  22.093099  10.597930  23.4000967  24.135124   4.3990460   1.6725421
#> 12    11  17.173040  19.538123  28.9686687  30.324161  -2.0493732  -6.4974194
#> 13    12  18.155999  24.452764  30.4697773  29.285293  -6.2057797  -5.7472217
#> 14    13  19.335205  30.018862  34.7096864  30.368435  -2.8743037  -5.9573272
#> 15    14  21.297583  31.008264  36.4160837  28.106458   0.6650098  -0.6239205
#> 16    15  25.147936  35.890228  38.3318462  29.986441   8.8944767   5.3669748
#> 17    16  33.174406  37.205731  41.9079809  31.655307  12.2870129   7.6571515
#> 18    17  36.627511  40.976212  43.7333457  36.390592  17.4806309   7.6973495
#> 19    18  43.583445  40.416869  43.9912148  34.786635  19.3823539  13.2387687
#> 20    19  49.667227  41.037287  49.2081509  33.109245  26.3333552   8.9036159
#> 21    20  51.962827  48.048207  51.1814373  38.364500  21.3731996  10.1933913
#> 22    21  51.531570  51.788361  54.5823268  41.608966  20.1121208  12.5172603
#> 23    22  49.991255  53.019131  55.5900083  45.813145  21.4332783  13.9358866
#> 24    23  52.870272  55.338751  58.4333418  50.292744  28.8056540  18.9842139
#> 25    24  51.690341  56.141628  59.2617662  49.737910  31.3963320  21.1028994
#> 26    25  54.293777  59.923113  61.1796899  56.211703  37.0895157  22.5777495
#> 27    26  54.882130  61.268241  63.8673783  59.800226  38.3779505  25.0833907
#> 28    27  58.026693  63.476944  66.9474492  62.797131  39.3762262  26.7205496
#> 29    28  60.250559  63.118474  70.6555200  61.693772  40.3698473  27.5409290
#> 30    29  62.212021  62.099662  69.9915350  59.492379  45.0715457  30.6715306
#> 31    30  62.253138  65.295938  71.2797911  62.394156  47.7762233  35.5110452
#> 32    31  59.914474  67.571199  74.0808197  64.875843  51.0918572  35.8482494
#> 33    32  59.308245  63.725516  76.6011959  65.506807  53.0942325  33.3666959
#> 34    33  61.707895  61.909034  80.9569478  65.579961  59.0070408  37.6426573
#> 35    34  63.622108  61.383099  81.5119062  62.768727  59.0966462  35.5710539
#> 36    35  71.958784  68.337551  87.8733614  61.744223  65.8306168  33.8072336
#> 37    36  74.876788  69.084328  85.1196322  56.425975  65.8542388  39.9168746
#> 38    37  74.037355  69.906445  87.5973141  62.466719  61.3980453  40.5106907
#> 39    38  71.823718  70.706557  90.8581214  60.210831  63.8407170  42.1397537
#> 40    39  72.562378  65.210274  98.5701890  63.310248  68.5259331  47.4801291
#> 41    40  81.492862  70.264318  97.9596211  69.775021  71.2582866  54.4485922
#> 42    41  78.048599  70.835804 102.4885512  71.370902  71.6489852  60.2695332
#> 43    42  83.821235  74.048568 105.1066917  71.597486  78.7679942  60.9881838
#> 44    43  85.245253  78.153936 109.6711754  74.554180  76.5856902  57.0302668
#> 45    44  89.862399  84.735838 115.8848110  76.659354  84.6023878  60.1720645
#> 46    45  92.663922  83.409647 114.8338234  83.274549  87.1344266  66.1480042
#> 47    46  96.187395  80.651583 117.5812156  84.683325  89.6667719  65.6516884
#> 48    47 100.523533  77.484812 121.1034428  82.581947  95.2177426  68.8983403
#> 49    48 100.383699  78.629970 126.4542537  85.200337  91.7986834  68.1954016
#> 50    49 101.442392  81.599832 124.9403018  90.305188  93.2566904  71.9014295
#> 51    50 104.531305  81.581045 128.3329791  88.395995 100.2256917  78.5283602
#> 52    51 105.706899  84.638175 129.3055130  89.244137 103.1141276  84.1120359
#> 53    52 103.919187  88.686313 134.0494079  91.926881  99.4880922  81.4154990
#> 54    53 105.910466  88.730958 138.4771706  96.954681 103.3608957  88.6792016
#> 55    54 108.151788  89.055969 140.7393678 102.130371 103.1589168  91.5265203
#> 56    55 113.878242  95.784992 142.7995404 108.208127 102.4677905  94.8640022
#> 57    56 114.958008  99.750755 148.1047941 111.055993 106.4063360  99.2455724
#> 58    57 117.442082 104.693893 148.5569502 112.685756 106.3716915 101.8413993
#> 59    58 116.950378 107.454041 145.8825676 112.835955 112.5152768 103.9489769
#> 60    59 120.704653 105.684445 150.1737218 112.324260 117.9455416 106.9772696
#> 61    60 121.629772 107.056457 154.8921156 109.810007 118.2155587 106.8380658
#> 62    61 125.720598 107.052244 155.6415832 112.831976 120.0993301 108.9687253
#> 63    62 136.177711 105.994713 160.7300790 115.745435 120.9489838 107.2640588
#> 64    63 136.461314 110.301257 165.8169650 121.413228 121.4924628 108.7292212
#> 65    64 141.180229 109.706124 164.6275095 123.733045 122.4856852 108.5334796
#> 66    65 142.652609 112.247986 167.2877261 122.961669 122.5734355 106.9163278
#> 67    66 142.511534 116.560604 174.7690839 128.596316 123.0291980 110.7718778
#> 68    67 148.567634 111.747669 170.6788888 129.514048 122.2855255 109.4471442
#> 69    68 150.179388 110.289212 173.2140417 129.468271 123.3494478 107.1898593
#> 70    69 152.531158 114.746896 173.7205813 133.313219 126.2663947 108.3164736
#> 71    70 151.074606 110.260596 172.1499950 134.663826 128.6590135 108.5997916
#> 72    71 152.865593 112.316619 173.9423880 134.440376 131.6045584 115.7139021
#> 73    72 154.555921 111.685431 185.0267524 133.389580 132.6342044 116.0633398
#> 74    73 156.924453 108.297988 187.0644694 137.542040 132.5032184 115.7492716
#> 75    74 154.109106 110.936206 191.6481756 138.456017 134.1435561 118.7878071
#> 76    75 156.837190 107.440948 197.4791786 141.852944 140.1665661 118.0945969
#> 77    76 158.982453 109.236890 207.8417032 151.042538 141.2191274 116.1753148
#> 78    77 161.422585 109.808806 207.0480661 156.945342 143.1710981 122.8229845
#> 79    78 170.502095 114.564310 205.3772001 156.867563 141.0051419 125.0007333
#> 80    79 178.618495 113.245778 210.0312163 159.705531 147.5022842 122.3791309
#> 81    80 179.505334 116.662769 212.6497188 163.561727 151.1767040 125.1122520
#> 82    81 179.829506 117.998316 209.4543991 164.837474 150.4986897 130.6157159
#> 83    82 180.860443 115.423231 213.1654904 167.333116 147.5676289 133.4426816
#> 84    83 189.038579 119.072040 212.6884025 167.096394 147.6512064 135.7072731
#> 85    84 192.757613 117.049036 213.2144685 161.597018 147.7076093 135.8670484
#> 86    85 188.811288 118.716000 217.4775543 165.367927 150.5312703 135.3964916
#> 87    86 194.359800 122.221336 222.5057269 173.908087 156.3807226 136.4795066
#> 88    87 195.642128 126.573466 223.3941947 173.396111 160.3567241 141.5736483
#> 89    88 193.801905 132.141026 224.9771334 181.140163 162.4419583 139.1672954
#> 90    89 195.213883 132.423351 225.6526195 182.552214 164.6840016 143.2599990
#> 91    90 193.826963 130.038402 232.0127904 181.917187 166.0897573 150.3493232
#> 92    91 194.068189 126.504681 233.3815619 188.435368 173.9874970 150.7363289
#> 93    92 198.920373 127.252382 236.9803711 187.482569 171.6210965 152.0451894
#> 94    93 198.989399 130.017191 233.0845378 189.695375 177.9902434 153.8155686
#> 95    94 203.066080 135.147253 232.8976086 198.014828 183.6395652 150.4864027
#> 96    95 210.546069 135.497371 235.8625486 203.035446 186.9301580 155.2448351
#> 97    96 209.044820 138.163329 242.5353465 204.947762 186.9291115 157.9239518
#> 98    97 211.725819 143.226332 244.8330655 201.552458 189.4275333 156.7291668
#> 99    98 218.968281 148.586773 248.5243448 208.262525 190.9736314 157.2599939
#> 100   99 222.268775 148.320567 249.1088410 207.667157 192.5288870 160.4305295
#> 101  100 228.534494 150.124417 252.1546504 212.522512 195.8694439 162.3586054
#> 102  101 229.890879 151.097115 253.7951878 216.621981 196.3204614 157.8903782
#> 103  102 233.362730 157.063536 254.7286119 218.878395 196.2892169 155.9264679
#> 104  103 232.935976 162.844467 257.6639339 222.695518 196.9448946 159.8685667
#> 105  104 231.852916 164.860962 259.6342063 230.687923 198.0313558 167.1260970
#> 106  105 235.461266 170.216157 259.1145868 233.925249 204.5683949 165.9545837
#> 107  106 232.435119 170.291518 259.9300295 233.959626 209.9610145 167.6149521
#> 108  107 233.161186 174.105761 264.6147814 234.437884 219.5818658 165.5942493
#> 109  108 233.850002 177.789782 265.6067521 238.431577 222.7891403 169.7864057
#> 110  109 234.215957 179.414554 267.9485765 239.575650 225.8121127 171.8473835
#> 111  110 233.198390 181.792608 272.6547125 245.867636 226.2940896 178.0579810
#> 112  111 234.521443 182.473567 271.5739338 247.849744 227.2624817 181.8858352
#> 113  112 233.519133 186.167737 273.7624823 251.552931 231.3710343 186.7230723
#> 114  113 236.642389 189.689591 274.8128346 255.704941 236.4539514 189.9489576
#> 115  114 239.220079 192.760043 277.7398772 256.130489 234.0915838 189.9187893
#> 116  115 246.071963 199.040656 274.7205063 262.600779 235.4245160 193.7348418
#> 117  116 246.161861 202.716035 271.9775959 271.917295 235.2244280 195.3282148
#> 118  117 249.632174 210.662319 279.4762682 276.344439 239.6576402 199.9788970
#> 119  118 254.347906 212.719431 279.4355167 274.509868 245.3201132 196.9763875
#> 120  119 260.723239 211.392227 279.6745926 278.017447 242.5379449 197.9643311
#> 121  120 261.734149 213.864549 281.1511888 283.666434 248.6895085 199.0308195
#> 122  121 261.562851 217.151930 282.0916933 287.669557 247.9019704 198.5857928
#> 123  122 265.342431 217.982911 282.9936814 289.867784 251.3296480 204.0122313
#> 124  123 267.414127 220.554656 280.3340448 290.844776 250.4307228 208.1822857
#> 125  124 266.550905 227.415742 283.2467815 297.608783 251.1066865 211.6997847
#> 126  125 271.228902 232.396783 289.0262064 304.505942 254.6083275 218.8093335
#> 127  126 272.180593 237.150453 289.4661005 310.615415 256.5773321 221.5098238
#> 128  127 275.529190 239.701623 287.9342424 311.440490 259.6595286 223.5598696
#> 129  128 276.333427 244.836432 286.7433663 309.123193 260.1989415 222.9836374
#> 130  129 277.114882 243.644864 288.6056072 312.342918 267.8948841 223.1793443
#> 131  130 277.899778 243.872555 290.7362645 315.115185 268.0659106 231.8309573
#> 132  131 274.669689 246.966256 294.7493881 319.960003 267.3165054 230.4017349
#> 133  132 272.040059 249.381042 298.8932663 320.910472 266.2036634 234.2683921
#> 134  133 271.366267 252.793286 297.6714808 319.804537 264.0838615 237.2963864
#> 135  134 272.719076 259.444140 299.4886421 325.041894 266.4620748 238.4939286
#> 136  135 275.513896 261.150134 299.3966358 330.559807 269.1272463 240.6829686
#> 137  136 281.876225 265.943179 303.7765310 330.023998 272.2470147 239.2791604
#> 138  137 284.589792 270.830691 302.6817703 326.103900 271.6572386 238.2305485
#> 139  138 285.611964 274.119451 304.8132621 331.936957 273.9598137 237.1459976
#> 140  139 282.671498 273.436366 303.5820605 330.694172 272.4050925 241.0807796
#> 141  140 285.278167 271.186792 301.4742155 335.299347 272.5442976 242.7571904
#> 142  141 289.541180 274.751183 301.8717430 334.661302 274.0942473 245.5704253
#> 143  142 284.748923 278.627337 306.6696539 340.995321 279.4996475 247.2768081
#> 144  143 290.471674 281.203314 310.6367149 347.106549 274.8371155 253.6577446
#> 145  144 296.024633 276.482881 313.8795689 344.751500 272.9767187 256.0105715
#> 146  145 297.807870 276.663311 315.0251207 351.533975 277.0395210 260.7734559
#> 147  146 298.794462 281.763496 317.1329301 356.807117 280.8032043 262.1553504
#> 148  147 297.976779 283.019512 315.2379037 356.072180 276.9766306 264.0585509
#> 149  148 301.790968 287.510487 313.4594714 363.027644 278.0054573 265.9626294
#> 150  149 305.500331 288.816117 314.4377225 369.789485 274.9581046 262.6996520
#> 151  150 308.133226 287.637186 312.8198024 369.843654 276.9095591 260.9356385
#> 152  151 314.063195 288.974580 317.0103803 369.850989 276.7098615 261.3365437
#> 153  152 318.437825 287.305755 316.3829437 365.998855 277.9839148 260.5035527
#> 154  153 319.839745 289.344762 320.3731837 368.652337 281.1703406 260.4393035
#> 155  154 318.491256 292.516131 320.6694616 371.885827 285.6533801 263.1312999
#> 156  155 318.489893 294.498196 319.5272430 370.840515 286.1693583 268.9708920
#> 157  156 319.104188 300.101441 324.5099045 375.083675 287.8371085 269.0723006
#> 158  157 323.138436 302.498832 325.1750361 374.020853 289.8174816 271.3438105
#> 159  158 327.478242 303.119260 325.0460033 376.475516 292.5905745 269.5709627
#> 160  159 327.507956 303.215005 323.2352164 377.739437 288.7805778 271.0827589
#> 161  160 330.588418 305.983759 327.2946838 376.745180 293.1464447 273.7313549
#> 162  161 328.728792 312.809824 331.6228955 382.902499 296.2309407 268.0254012
#> 163  162 325.691138 315.023011 337.8985744 378.725274 302.2719079 275.2618125
#> 164  163 328.495334 317.812470 337.8601230 377.524398 302.1681250 273.7590118
#> 165  164 332.170492 321.558678 336.8400056 379.816097 305.2699279 280.3735717
#> 166  165 335.628238 321.935979 337.1901218 380.262844 304.7328215 273.8841802
#> 167  166 333.926686 322.822259 337.6376277 386.280823 304.2606337 279.4204890
#> 168  167 334.194742 323.016099 342.0577048 388.603675 299.5378224 280.4550743
#> 169  168 333.119890 319.343057 339.6475792 389.462670 299.3809168 281.6675545
#> 170  169 332.344584 321.555302 348.4859963 394.213705 305.3012284 287.2880066
#> 171  170 339.024215 322.869941 351.0949627 394.381337 305.5465817 284.9519287
#> 172  171 337.863784 327.378982 349.7816616 397.676756 314.3512780 281.0269778
#> 173  172 339.265370 329.335594 351.2201882 397.640601 317.5036072 279.7955332
#> 174  173 342.040578 332.260152 357.1223056 407.330387 321.4279537 279.0401112
#> 175  174 344.884152 332.833992 359.0057465 411.129930 324.4414078 276.7525924
#> 176  175 349.967036 336.295335 363.3975349 416.998031 320.6045096 280.8433682
#> 177  176 347.661207 341.890721 364.7124835 423.030938 317.2391933 280.1218147
#> 178  177 351.030417 341.835377 367.4502848 426.027193 317.4629442 286.9412222
#> 179  178 355.697648 341.483033 366.5727157 437.439638 320.7544806 292.4635112
#> 180  179 353.822684 349.343226 367.1562986 440.713589 321.4302949 300.7981250
#> 181  180 361.702966 353.604405 376.2695936 446.362454 326.9502808 307.3487119
#> 182  181 363.777701 354.074205 380.6156179 450.020029 329.9374880 310.1246705
#> 183  182 364.800090 350.177086 380.7751922 452.101955 335.4064387 311.0644740
#> 184  183 362.955599 351.593305 383.0034065 460.828175 336.8479474 321.3195957
#> 185  184 368.060494 352.743115 381.8498738 466.399289 333.8453838 329.4363816
#> 186  185 366.263541 354.627137 382.6256037 472.251410 335.0867963 333.7685661
#> 187  186 367.830642 360.150216 382.0179143 473.550224 337.0700673 335.9853436
#> 188  187 364.388493 363.177800 382.5814372 473.527469 341.0834834 340.2904622
#> 189  188 369.599296 362.419610 382.5384356 472.801786 340.9279551 351.4699394
#> 190  189 375.404746 357.803318 385.7517290 469.372824 343.0546933 349.8888396
#> 191  190 377.356136 360.336461 390.1809120 472.270616 345.3650430 356.9165054
#> 192  191 380.474850 365.920792 388.6813993 480.137739 346.8965789 354.1610440
#> 193  192 381.986905 365.112169 396.0408974 483.038887 346.2457669 359.0955187
#> 194  193 382.174296 357.771719 397.0374795 484.999944 344.3368102 364.5014679
#> 195  194 382.082534 363.392620 399.3506954 488.559148 340.6387508 367.8735516
#> 196  195 382.626247 362.261174 402.1860974 489.268981 346.4661264 368.2389880
#> 197  196 387.035021 366.345780 409.2073353 493.320198 350.7362020 369.0263370
#> 198  197 388.966460 369.166540 404.1986046 495.403737 344.3088119 370.7642779
#> 199  198 388.279145 371.898797 407.2578432 500.797087 353.4324983 372.0168101
#> 200  199 388.856631 377.659182 411.1875837 499.296226 360.5645659 375.7288643
#> 
#> [[2]]
#>     Time            1            2           3            4           5
#> 1      0  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> 2      1  4.722416520 -3.564217782  0.64123216  4.839178966 -1.07221587
#> 3      2  7.713235191 -0.458614831  3.14443202  2.187538077  0.56848033
#> 4      3  2.103860419 -0.195179400  1.73759801  3.519961177  1.10396182
#> 5      4 -1.748996942  0.299041266 -0.15703660  6.356575108  2.57326905
#> 6      5  4.263031398 -1.873103672  2.36681364 -0.089497620  2.77293314
#> 7      6 -1.824999455  4.243427671  2.19741412 -1.326498483 -0.28066354
#> 8      7  4.076824099  2.814650037  3.31360946  3.912440067  0.44709422
#> 9      8  0.988822794  4.942110395  1.90080854  0.447432252 -0.71888122
#> 10     9  0.871882551  3.251114606  1.78967310  8.236297855 -4.98826354
#> 11    10  0.927022125  1.138701386  6.46555226 -3.948303399  3.99333163
#> 12    11 -4.920058377  8.940193651  5.56857199  6.189037205 -6.44841919
#> 13    12  0.982959081  4.914640455  1.50110856 -1.038868551 -4.15640655
#> 14    13  1.179205813  5.566098624  4.23990910  1.083142669  3.33147600
#> 15    14  1.962377731  0.989401649  1.70639732 -2.261976957  3.53931349
#> 16    15  3.850352896  4.881963991  1.91576249  1.879982175  8.22946697
#> 17    16  8.026469659  1.315502469  3.57613476  1.668866661  3.39253613
#> 18    17  3.453105327  3.770481650  1.82536475  4.735285115  5.19361800
#> 19    18  6.955933868 -0.559343140  0.25786908 -1.603956953  1.90172304
#> 20    19  6.083782114  0.620418289  5.21693614 -1.677390637  6.95100134
#> 21    20  2.295600207  7.010919356  1.97328637  5.255255285 -4.96015565
#> 22    21 -0.431257344  3.740154204  3.40088956  3.244466233 -1.26107879
#> 23    22 -1.540314999  1.230770203  1.00768146  4.204178620  1.32115748
#> 24    23  2.879017191  2.319620418  2.84333349  4.479598668  7.37237576
#> 25    24 -1.179931000  0.802876391  0.82842445 -0.554833738  2.59067793
#> 26    25  2.603436101  3.781485564  1.91792362  6.473792841  5.69318378
#> 27    26  0.588353078  1.345127619  2.68768841  3.588522903  1.28843479
#> 28    27  3.144562902  2.208703256  3.08007091  2.996905084  0.99827572
#> 29    28  2.223865716 -0.358469986  3.70807083 -1.103358956  0.99362103
#> 30    29  1.961462775 -1.018812747 -0.66398497 -2.201392390  4.70169842
#> 31    30  0.041116831  3.196276392  1.28825608  2.901776877  2.70467757
#> 32    31 -2.338664602  2.275260670  2.80102856  2.481687319  3.31563392
#> 33    32 -0.606228395 -3.845683068  2.52037624  0.630963561  2.00237528
#> 34    33  2.399650002 -1.816482067  4.35575191  0.073154094  5.91280836
#> 35    34  1.914212754 -0.525934266  0.55495834 -2.811234621  0.08960542
#> 36    35  8.336676135  6.954451367  6.36145520 -1.024503510  6.73397052
#> 37    36  2.918004013  0.746776947 -2.75372918 -5.318247750  0.02362208
#> 38    37 -0.839433661  0.822117766  2.47768187  6.040743919 -4.45619355
#> 39    38 -2.213636234  0.800111477  3.26080734 -2.255888404  2.44267169
#> 40    39  0.738659890 -5.496282329  7.71206763  3.099416944  4.68521611
#> 41    40  8.930483889  5.054043585 -0.61056788  6.464773668  2.73235348
#> 42    41 -3.444262779  0.571485474  4.52893009  1.595880667  0.39069863
#> 43    42  5.772635562  3.212764078  2.61814050  0.226583745  7.11900904
#> 44    43  1.424018557  4.105368693  4.56448367  2.956694213 -2.18230401
#> 45    44  4.617145513  6.581901310  6.21363555  2.105173658  8.01669760
#> 46    45  2.801523473 -1.326191081 -1.05098751  6.615195116  2.53203874
#> 47    46  3.523472909 -2.758063981  2.74739219  1.408776065  2.53234535
#> 48    47  4.336137898 -3.166771006  3.52222721 -2.101377559  5.55097068
#> 49    48 -0.139833770  1.145158781  5.35081086  2.618389975 -3.41905923
#> 50    49  1.058692530  2.969861337 -1.51395189  5.104851219  1.45800701
#> 51    50  3.088913498 -0.018786956  3.39267728 -1.909193000  6.96900134
#> 52    51  1.175593950  3.057130332  0.97253385  0.848141973  2.88843590
#> 53    52 -1.787712463  4.048137940  4.74389499  2.682743111 -3.62603546
#> 54    53  1.991278586  0.044645264  4.42776264  5.027800258  3.87280352
#> 55    54  2.241322801  0.325010542  2.26219725  5.175690627 -0.20197889
#> 56    55  5.726454130  6.729023054  2.06017258  6.077755069 -0.69112625
#> 57    56  1.079766033  3.965763617  5.30525367  2.847866979  3.93854542
#> 58    57  2.484073837  4.943137520  0.45215613  1.629762299 -0.03464441
#> 59    58 -0.491704185  2.760148029 -2.67438265  0.150199503  6.14358529
#> 60    59  3.754275261 -1.769596009  4.29115424 -0.511694809  5.43026481
#> 61    60  0.925118630  1.372012435  4.71839377 -2.514253059  0.27001701
#> 62    61  4.090826029 -0.004213572  0.74946757  3.021968375  1.88377140
#> 63    62 10.457112650 -1.057531336  5.08849585  2.913458895  0.84965375
#> 64    63  0.283603423  4.306544006  5.08688603  5.667793775  0.54347901
#> 65    64  4.718914427 -0.595132324 -1.18945553  2.319816204  0.99322242
#> 66    65  1.472380803  2.541861317  2.66021664 -0.771375723  0.08775025
#> 67    66 -0.141075720  4.312618294  7.48135780  5.634646688  0.45576253
#> 68    67  6.056100743 -4.812934733 -4.09019511  0.917732294 -0.74367257
#> 69    68  1.611754043 -1.458456923  2.53515287 -0.045776895  1.06392232
#> 70    69  2.351769172  4.457683595  0.50653962  3.844948078  2.91694695
#> 71    70 -1.456551616 -4.486299626 -1.57058633  1.350606587  2.39261879
#> 72    71  1.790987066  2.056023077  1.79239296 -0.223449834  2.94554487
#> 73    72  1.690327945 -0.631188444 11.08436447 -1.050795530  1.02964603
#> 74    73  2.368531694 -3.387442481  2.03771693  4.152459311 -0.13098599
#> 75    74 -2.815347150  2.638218206  4.58370619  0.913977357  1.64033770
#> 76    75  2.728084492 -3.495258938  5.83100305  3.396926662  6.02300999
#> 77    76  2.145263278  1.795942137 10.36252464  9.189594797  1.05256134
#> 78    77  2.440131210  0.571915999 -0.79363712  5.902804007  1.95197067
#> 79    78  9.079510879  4.755504138 -1.67086599 -0.077779736 -2.16595618
#> 80    79  8.116399467 -1.318531445  4.65401613  2.837968608  6.49714228
#> 81    80  0.886839206  3.416991001  2.61850258  3.856195775  3.67441982
#> 82    81  0.324171745  1.335546778 -3.19531978  1.275746950 -0.67801431
#> 83    82  1.030937269 -2.575084898  3.71109138  2.495641789 -2.93106086
#> 84    83  8.178135419  3.648808762 -0.47708792 -0.236721740  0.08357751
#> 85    84  3.719034087 -2.023004009  0.52606603 -5.499375868  0.05640298
#> 86    85 -3.946324266  1.666964318  4.26308580  3.770908972  2.82366091
#> 87    86  5.548511254  3.505336012  5.02817252  8.540160113  5.84945231
#> 88    87  1.282328428  4.352129672  0.88846779 -0.511975862  3.97600155
#> 89    88 -1.840222655  5.567560367  1.58293874  7.744051163  2.08523422
#> 90    89  1.411977575  0.282325001  0.67548614  1.412051360  2.24204325
#> 91    90 -1.386919967 -2.384949817  6.36017085 -0.635026724  1.40575568
#> 92    91  0.241226170 -3.533720089  1.36877154  6.518180876  7.89773977
#> 93    92  4.852184099  0.747700825  3.59880916 -0.952799302 -2.36640049
#> 94    93  0.069025322  2.764808375 -3.89583331  2.212806621  6.36914690
#> 95    94  4.076681393  5.130062386 -0.18692913  8.319452958  5.64932178
#> 96    95  7.479988761  0.350118386  2.96493997  5.020617348  3.29059272
#> 97    96 -1.501248695  2.665957122  6.67279790  1.912316682 -0.00104645
#> 98    97  2.680998844  5.063003256  2.29771903 -3.395304304  2.49842176
#> 99    98  7.242462435  5.360441550  3.69127927  6.710066908  1.54609817
#> 100   99  3.300494152 -0.266206062  0.58449618 -0.595368136  1.55525555
#> 101  100  6.265718111  1.803849852  3.04580946  4.855355466  3.34055688
#> 102  101  1.356385575  0.972698203  1.64053737  4.099468572  0.45101749
#> 103  102  3.471850824  5.966421033  0.93342415  2.256413925 -0.03124443
#> 104  103 -0.426753865  5.780930090  2.93532198  3.817123237  0.65567767
#> 105  104 -1.083060217  2.016495009  1.97027241  7.992404591  1.08646122
#> 106  105  3.608350620  5.355195027 -0.51961957  3.237326690  6.53703903
#> 107  106 -3.026147155  0.075361916  0.81544275  0.034376203  5.39261964
#> 108  107  0.726066985  3.814242559  4.68475188  0.478258009  9.62085134
#> 109  108  0.688816044  3.684021396  0.99197073  3.993693866  3.20727446
#> 110  109  0.365954907  1.624772046  2.34182437  1.144072329  3.02297244
#> 111  110 -1.017567024  2.378053820  4.70613604  6.291986299  0.48197691
#> 112  111  1.323052856  0.680958890 -1.08077872  1.982108057  0.96839205
#> 113  112 -1.002309666  3.694170272  2.18854847  3.703186798  4.10855258
#> 114  113  3.123256006  3.521853664  1.05035229  4.152010247  5.08291714
#> 115  114  2.577689073  3.070452141  2.92704263  0.425547654 -2.36236761
#> 116  115  6.851884510  6.280613150 -3.01937096  6.470290370  1.33293217
#> 117  116  0.089898199  3.675378477 -2.74291036  9.316515564 -0.20008797
#> 118  117  3.470312502  7.946283806  7.49867227  4.427144074  4.43321220
#> 119  118  4.715731851  2.057112388 -0.04075151 -1.834571076  5.66247305
#> 120  119  6.375333113 -1.327204205  0.23907592  3.507578833 -2.78216832
#> 121  120  1.010910787  2.472322373  1.47659623  5.648987511  6.15156358
#> 122  121 -0.171298521  3.287381147  0.94050445  4.003123062 -0.78753811
#> 123  122  3.779580394  0.830980243  0.90198813  2.198226615  3.42767764
#> 124  123  2.071695279  2.571744962 -2.65963659  0.976992511 -0.89892523
#> 125  124 -0.863221914  6.861085978  2.91273668  6.764006510  0.67596365
#> 126  125  4.677997577  4.981041126  5.77942489  6.897159546  3.50164103
#> 127  126  0.951690797  4.753669858  0.43989409  6.109472480  1.96900457
#> 128  127  3.348597068  2.551170323 -1.53185803  0.825075426  3.08219656
#> 129  128  0.804236613  5.134809383 -1.19087616 -2.317297025  0.53941289
#> 130  129  0.781455263 -1.191568208  1.86224089  3.219724522  7.69594262
#> 131  130  0.784895563  0.227690584  2.13065733  2.772267441  0.17102645
#> 132  131 -3.230088773  3.093701808  4.01312363  4.844818256 -0.74940519
#> 133  132 -2.629630318  2.414785294  4.14387818  0.950468679 -1.11284200
#> 134  133 -0.673791952  3.412244133 -1.22178546 -1.105935649 -2.11980185
#> 135  134  1.352809580  6.650854563  1.81716121  5.237357193  2.37821327
#> 136  135  2.794819930  1.705993816 -0.09200627  5.517913444  2.66517149
#> 137  136  6.362329081  4.793045155  4.37989522 -0.535809005  3.11976838
#> 138  137  2.713567171  4.887511987 -1.09476073 -3.920098047 -0.58977607
#> 139  138  1.022171424  3.288759581  2.13149178  5.833057113  2.30257509
#> 140  139 -2.940465940 -0.683085010 -1.23120161 -1.242785412 -1.55472123
#> 141  140  2.606669123 -2.249574374 -2.10784493  4.605174809  0.13920508
#> 142  141  4.263012591  3.564391841  0.39752752 -0.638044378  1.54994970
#> 143  142 -4.792256245  3.876153543  4.79791089  6.334018372  5.40540027
#> 144  143  5.722751029  2.575976801  3.96706099  6.111228822 -4.66253205
#> 145  144  5.552959030 -4.720433178  3.24285397 -2.355049853 -1.86039680
#> 146  145  1.783236723  0.180430809  1.14555179  6.782475673  4.06280231
#> 147  146  0.986592381  5.100184775  2.10780939  5.273142197  3.76368325
#> 148  147 -0.817683621  1.256016116 -1.89502642 -0.734937042 -3.82657362
#> 149  148  3.814189491  4.490974969 -1.77843230  6.955463786  1.02882662
#> 150  149  3.709363025  1.305629361  0.97825118  6.761840450 -3.04735263
#> 151  150  2.632894724 -1.178930128 -1.61792013  0.054169709  1.95145447
#> 152  151  5.929968557  1.337393299  4.19057793  0.007334508 -0.19969760
#> 153  152  4.374630190 -1.668824817 -0.62743663 -3.852134231  1.27405331
#> 154  153  1.401920262  2.039007498  3.99024003  2.653482861  3.18642573
#> 155  154 -1.348489229  3.171368033  0.29627780  3.233489769  4.48303953
#> 156  155 -0.001362768  1.982065941 -1.14221855 -1.045311917  0.51597817
#> 157  156  0.614294674  5.603244525  4.98266148  4.243159941  1.66775025
#> 158  157  4.034248701  2.397391384  0.66513160 -1.062822771  1.98037306
#> 159  158  4.339805420  0.620427948 -0.12903279  2.454663674  2.77309297
#> 160  159  0.029714455  0.095744248 -1.81078692  1.263920647 -3.80999678
#> 161  160  3.080461299  2.768754009  4.05946741 -0.994257181  4.36586694
#> 162  161 -1.859625491  6.826065444  4.32821170  6.157319667  3.08449605
#> 163  162 -3.037653739  2.213186956  6.27567894 -4.177225249  6.04096712
#> 164  163  2.804195537  2.789459310 -0.03845143 -1.200876541 -0.10378290
#> 165  164  3.675158032  3.746208087 -1.02011736  2.291699755  3.10180294
#> 166  165  3.457745996  0.377300315  0.35011615  0.446746481 -0.53710637
#> 167  166 -1.701552373  0.886280785  0.44750597  6.017978955 -0.47218784
#> 168  167  0.268055948  0.193839681  4.42007703  2.322852156 -4.72281128
#> 169  168 -1.074851555 -3.673042253 -2.41012558  0.858995503 -0.15690564
#> 170  169 -0.775305491  2.212245467  8.83841716  4.751035009  5.92031161
#> 171  170  6.679630394  1.314639119  2.60896634  0.167631812  0.24535331
#> 172  171 -1.160430474  4.509040940 -1.31330108  3.295418588  8.80469634
#> 173  172  1.401585497  1.956611878  1.43852655 -0.036154934  3.15232919
#> 174  173  2.775208290  2.924557695  5.90211740  9.689785795  3.92434651
#> 175  174  2.843574079  0.573839661  1.88344092  3.799543785  3.01345411
#> 176  175  5.082883917  3.461343400  4.39178839  5.868100224 -3.83689825
#> 177  176 -2.305829390  5.595385837  1.31494858  6.032907795 -3.36531626
#> 178  177  3.369210335 -0.055344062  2.73780132  2.996254739  0.22375084
#> 179  178  4.667231011 -0.352343717 -0.87756908 11.412444441  3.29153643
#> 180  179 -1.874964570  7.860193335  0.58358287  3.273951799  0.67581429
#> 181  180  7.880282677  4.261178656  9.11329508  5.648864578  5.51998595
#> 182  181  2.074734791  0.469800065  4.34602427  3.657575125  2.98720720
#> 183  182  1.022388891 -3.897119113  0.15957428  2.081925977  5.46895070
#> 184  183 -1.844491289  1.416219073  2.22821433  8.726219672  1.44150869
#> 185  184  5.104895321  1.149809748 -1.15353276  5.571114413 -3.00256367
#> 186  185 -1.796952519  1.884022586  0.77572998  5.852120664  1.24141255
#> 187  186  1.567100226  5.523078367 -0.60768948  1.298814493  1.98327095
#> 188  187 -3.442148224  3.027584110  0.56352290 -0.022755559  4.01341614
#> 189  188  5.210802967 -0.758189634 -0.04300154 -0.725682337 -0.15552827
#> 190  189  5.805449948 -4.616291866  3.21329340 -3.428962639  2.12673820
#> 191  190  1.951389704  2.533143038  4.42918296  2.897791693  2.31034971
#> 192  191  3.118714272  5.584330483 -1.49951272  7.867123716  1.53153587
#> 193  192  1.512054451 -0.808623167  7.35949817  2.901147559 -0.65081205
#> 194  193  0.187391359 -7.340449301  0.99658208  1.961056935 -1.90895669
#> 195  194 -0.091762616  5.620900540  2.31321585  3.559203923 -3.69805935
#> 196  195  0.543713898 -1.131445525  2.83540203  0.709833557  5.82737562
#> 197  196  4.408773607  4.084605807  7.02123789  4.051216569  4.27007560
#> 198  197  1.931438998  2.820759483 -5.00873067  2.083539261 -6.42739009
#> 199  198 -0.687314670  2.732257434  3.05923860  5.393350033  9.12368633
#> 200  199  0.577485290  5.760385027  3.92974052 -1.500861378  7.13206767
#>               6
#> 1    0.00000000
#> 2    2.61592633
#> 3   -1.72496543
#> 4   -1.79415024
#> 5   -4.98206109
#> 6    2.75730539
#> 7    1.45899307
#> 8    0.38382064
#> 9    1.95587523
#> 10   2.47866161
#> 11  -1.47686345
#> 12  -8.16996147
#> 13   0.75019770
#> 14  -0.21010547
#> 15   5.33340663
#> 16   5.99089534
#> 17   2.29017675
#> 18   0.04019795
#> 19   5.54141917
#> 20  -4.33515279
#> 21   1.28977540
#> 22   2.32386900
#> 23   1.41862635
#> 24   5.04832724
#> 25   2.11868556
#> 26   1.47485014
#> 27   2.50564117
#> 28   1.63715890
#> 29   0.82037942
#> 30   3.13060157
#> 31   4.83951464
#> 32   0.33720415
#> 33  -2.48155350
#> 34   4.27596141
#> 35  -2.07160340
#> 36  -1.76382032
#> 37   6.10964105
#> 38   0.59381609
#> 39   1.62906296
#> 40   5.34037541
#> 41   6.96846309
#> 42   5.82094105
#> 43   0.71865052
#> 44  -3.95791693
#> 45   3.14179769
#> 46   5.97593964
#> 47  -0.49631581
#> 48   3.24665195
#> 49  -0.70293874
#> 50   3.70602789
#> 51   6.62693070
#> 52   5.58367572
#> 53  -2.69653688
#> 54   7.26370257
#> 55   2.84731869
#> 56   3.33748192
#> 57   4.38157022
#> 58   2.59582684
#> 59   2.10757761
#> 60   3.02829274
#> 61  -0.13920382
#> 62   2.13065957
#> 63  -1.70466659
#> 64   1.46516248
#> 65  -0.19574161
#> 66  -1.61715179
#> 67   3.85554996
#> 68  -1.32473361
#> 69  -2.25728486
#> 70   1.12661427
#> 71   0.28331803
#> 72   7.11411045
#> 73   0.34943770
#> 74  -0.31406815
#> 75   3.03853547
#> 76  -0.69321018
#> 77  -1.91928208
#> 78   6.64766967
#> 79   2.17774882
#> 80  -2.62160244
#> 81   2.73312106
#> 82   5.50346394
#> 83   2.82696573
#> 84   2.26459148
#> 85   0.15977533
#> 86  -0.47055680
#> 87   1.08301500
#> 88   5.09414163
#> 89  -2.40635290
#> 90   4.09270365
#> 91   7.08932418
#> 92   0.38700571
#> 93   1.30886051
#> 94   1.77037921
#> 95  -3.32916589
#> 96   4.75843236
#> 97   2.67911673
#> 98  -1.19478499
#> 99   0.53082703
#> 100  3.17053566
#> 101  1.92807585
#> 102 -4.46822719
#> 103 -1.96391024
#> 104  3.94209877
#> 105  7.25753027
#> 106 -1.17151324
#> 107  1.66036837
#> 108 -2.02070282
#> 109  4.19215638
#> 110  2.06097783
#> 111  6.21059745
#> 112  3.82785428
#> 113  4.83723705
#> 114  3.22588536
#> 115 -0.03016837
#> 116  3.81605256
#> 117  1.59337294
#> 118  4.65068220
#> 119 -3.00250944
#> 120  0.98794355
#> 121  1.06648847
#> 122 -0.44502674
#> 123  5.42643847
#> 124  4.17005447
#> 125  3.51749892
#> 126  7.10954880
#> 127  2.70049029
#> 128  2.05004580
#> 129 -0.57623211
#> 130  0.19570684
#> 131  8.65161306
#> 132 -1.42922250
#> 133  3.86665726
#> 134  3.02799432
#> 135  1.19754217
#> 136  2.18903995
#> 137 -1.40380812
#> 138 -1.04861193
#> 139 -1.08455090
#> 140  3.93478202
#> 141  1.67641073
#> 142  2.81323491
#> 143  1.70638280
#> 144  6.38093652
#> 145  2.35282690
#> 146  4.76288438
#> 147  1.38189450
#> 148  1.90320049
#> 149  1.90407849
#> 150 -3.26297736
#> 151 -1.76401351
#> 152  0.40090524
#> 153 -0.83299104
#> 154 -0.06424914
#> 155  2.69199638
#> 156  5.83959208
#> 157  0.10140859
#> 158  2.27150991
#> 159 -1.77284779
#> 160  1.51179616
#> 161  2.64859607
#> 162 -5.70595374
#> 163  7.23641128
#> 164 -1.50280067
#> 165  6.61455985
#> 166 -6.48939147
#> 167  5.53630882
#> 168  1.03458534
#> 169  1.21248019
#> 170  5.62045203
#> 171 -2.33607782
#> 172 -3.92495095
#> 173 -1.23144463
#> 174 -0.75542198
#> 175 -2.28751877
#> 176  4.09077582
#> 177 -0.72155357
#> 178  6.81940755
#> 179  5.52228893
#> 180  8.33461383
#> 181  6.55058695
#> 182  2.77595859
#> 183  0.93980349
#> 184 10.25512171
#> 185  8.11678589
#> 186  4.33218448
#> 187  2.21677756
#> 188  4.30511852
#> 189 11.17947726
#> 190 -1.58109981
#> 191  7.02766577
#> 192 -2.75546139
#> 193  4.93447468
#> 194  5.40594928
#> 195  3.37208368
#> 196  0.36543637
#> 197  0.78734898
#> 198  1.73794097
#> 199  1.25253216
#> 200  3.71205424
#>